+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Upcoming version 0.26.0 (????-??-??)
 
-...
+### Changed
+
+* Separate plugin and theme update notification is sent for each available plugin and theme update [#167](https://github.com/chesio/bc-security/issues/167). Legacy behaviour can be enabled with following filters: `bc-security/filter:all-plugin-updates-in-one-notification` for plugins and `bc-security/filter:all-theme-updates-in-one-notification` for themes.
 
 ## Version 0.25.0 (2024-10-28)